+++ randomizer.py
@@ -508,6 +508,7 @@ class Randomizer:
     tweaks.remove_minor_panning_cutscenes(self)
     tweaks.fix_message_closing_sound_on_quest_status_screen(self)
     tweaks.fix_stone_head_bugs(self)
+    tweaks.show_number_of_tingle_statues_on_quest_status_screen(self)
     
     customizer.replace_link_model(self)
     tweaks.change_starting_clothes(self)